+++ examples/basic/README.md
@@ -2,4 +2,4 @@
 
 An end-to-end basic example that will provision the following:
 - A new resource group if one is not passed in.
-- A new Event Notification instance.
+- A new Event Notification lite instance.
